scraper/pyproject.toml
[project]
name = "scraper"
version = "0.1.0"
description = "Background worker for scraping products from different markets."
readme = "../README.md"
requires-python = ">=3.14"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "playwright>=1.58.0",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "ruff>=0.15.6",
]

[tool.ruff]
line-length = 88

target-version = "py314"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors (standard formatting/syntax errors)
    "F",  # pyflakes (unused variables, undefined names - catches real bugs!)
    "I",  # isort (automatically sorts and groups your imports alphabetically)
    "UP", # pyupgrade (automatically upgrades older Python syntax to modern 3.14 syntax)
]

ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["test"]
asyncio_mode = "auto"
